pyproject.toml
[tool.poetry]
name = "proteinworkshop"
version = "0.2.5"
description = ""
authors = ["Arian Jamasb <arian@jamasb.io>"]
readme = "README.md"
keywords = ["proteins", "deep learning", "datasets", "benchmarks"]
homepage = "https://www.proteins.sh"
license = "MIT"
repository = "https://www.github.com/a-r-j/ProteinWorkshop"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering"
]
packages = [
    { include = "proteinworkshop" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/a-r-j/ProteinWorkshop/issues"
"Documentation" = "https://proteins.sh"

[tool.poetry.dependencies]
python = ">= 3.9, < 3.11"
graphein = "^1.7.5"
loguru = "^0.7.0"
fair-esm = "^2.0.0"
hydra-core = "^1.3.2"
biotite = "^0.37.0"
e3nn = "^0.5.1"
einops = "^0.6.1"
beartype = "^0.15.0"
rich = "^13.5.2"
pytdc = "^0.4.1"
wandb = "^0.15.8"
torchmetrics = "^1.2.0"
lovely-tensors = "^0.1.15"
psutil = "^5.9.5"
tqdm = "^4.66.1"
jaxtyping = "^0.2.24"
omegaconf = "^2.3.0"
pytorch-lightning = "^2.0.7"
lightning = "^2.0.7"
python-dotenv = "^1.0.0"
wget = "^3.2"
opt-einsum = "^3.3.0"
torch = {version="^2.1.2", source="torch"}
pyrootutils = "^1.0.4"
torch-geometric = "^2.4.0"
hydra-colorlog = "^1.2.0"
foldcomp = "^0.0.5"
captum = "^0.6.0"
umap-learn = "^0.5.4"
matplotlib = "^3.7.2"
datashader = "^0.16.0"
holoviews = "^1.18.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
poethepoet = "^0.22.0"
ruff = "^0.1.1"
pydocstyle = "^6.3.0"
pre-commit = "^3.5.0"


[tool.poetry.group.test.dependencies]
pytest = "*"

[tool.poetry.group.docs.dependencies]
ipython = "^8.14.0"
furo = "^2023.8.19"
m2r2 = "^0.3.3.post2"
nbsphinx = "^0.9.2"
nbsphinx-link = "^1.3.0"
nbstripout = "^0.6.1"
pandoc = "^2.3"
pydocstyle = "^6.3.0"
sphinx = "^7.2.3"
sphinx-copybutton = "^0.5.2"
sphinx-inline-tabs = "^2023.4.21"
sphinxext-opengraph = "^0.8.2"
sphinxcontrib-gtagjs = "^0.2.1"
sphinxcontrib-jquery = "^4.1"
watermark = "^2.4.3"
sphinx-codeautolink = "^0.15.0"
emoji = "^2.8.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu/"
priority = "explicit"

[tool.poe.tasks]
install-torch-cluster = "pip install torch_cluster -f https://data.pyg.org/whl/torch-2.0.0+cpu.html"
install-torch-sparse = "pip install torch_sparse -f https://data.pyg.org/whl/torch-2.0.0+cpu.html"
install-torch-scatter = "pip install torch_scatter -f https://data.pyg.org/whl/torch-2.0.0+cpu.html"
install-torch-spline-conv = "pip install torch_spline_conv -f https://data.pyg.org/whl/torch-2.0.0+cpu.html"
install-openfold = "pip install git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769 && pip install git+https://github.com/aqlaboratory/openfold.git@4b41059694619831a7db195b7e0988fc4ff3a307"
install-torchdrug = "pip install torchdrug"
install-flash = "pip install lightning-flash --no-dependencies"

[tool.poetry.scripts]
train_workshop = "proteinworkshop.train:_main"
finetune_workshop = "proteinworkshop.finetune:_main"
install_pyg_workshop = "proteinworkshop.scripts.install_pyg:_install_pyg"
workshop = "proteinworkshop.scripts.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Code-formatting and code-style config
[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I"] # <-- Note: I suggest including the "I" for additional checks!
ignore = ["F722"] # Ignore Jaxtyping forward references: https://github.com/google/jaxtyping/blob/v0.2.12/FAQ.md

# Assuming you're developing for Python 3.10
target-version = "py310"

[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
  | .env
  | venv
  | .venv
)/
'''
line-length = 79
target-version = ['py310']

[tool.isort]
skip = [".env", "env", ".venv", "venv", "data"]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 79 # match our custom config above
multi_line_output = 3
float_to_top = true
use_parentheses = true
skip_glob = "**/*.pyx"
